refactor: separate account files in account folder
