+++ src/domain/usecases/account/add-account-usecase.spec.js
@@ -1,5 +1,3 @@
-const { MissingParamError } = require("../../utils/errors")
-const { UserAlreadyExistsError } = require("../errors")
 const AddAccountUseCase = require("./add-account-usecase")
 
 const makeFakeUserResult = () => ({
