+++ src/main/routes/account-router.js
@@ -3,7 +3,7 @@ const accountRouter = express.Router()
 
 const { adapt } = require('../adapters/express-router-adapter')
 
-const { makeAddAccountController } = require('../factory/controllers/add-account-controller-impl')
+const { makeAddAccountController } = require('../factory/controllers/account/add-account-controller-impl')
 
 accountRouter.post('/create', adapt(makeAddAccountController()))
